OCT-111: moved GetCategoriesByCodeQuery
