+++ components/catalogs/back/src/Infrastructure/Validation/ProductSelection/CategoriesCriterionContainsValidCategoriesValidator.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Akeneo\Catalogs\Infrastructure\Validation\ProductSelection;
 
-use Akeneo\Catalogs\Application\Persistence\GetCategoriesByCodeQueryInterface;
+use Akeneo\Catalogs\Application\Persistence\Category\GetCategoriesByCodeQueryInterface;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\ConstraintValidator;
 use Symfony\Component\Validator\Exception\UnexpectedTypeException;
