+++ components/catalogs/back/tests/Unit/Infrastructure/Controller/Internal/GetCategoriesActionTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Akeneo\Catalogs\Test\Unit\Infrastructure\Controller\Internal;
 
-use Akeneo\Catalogs\Application\Persistence\GetCategoriesByCodeQueryInterface;
+use Akeneo\Catalogs\Application\Persistence\Category\GetCategoriesByCodeQueryInterface;
 use Akeneo\Catalogs\Application\Persistence\GetCategoryTreeRootsQueryInterface;
 use Akeneo\Catalogs\Infrastructure\Controller\Internal\GetCategoriesAction;
 use PHPUnit\Framework\TestCase;
